Sémen a testy update
Už to skoro jede, boys

+++ test/test_examples/semantic_examples/semen_test_12_uncompatible_assignment_2.zig
@@ -5,7 +5,7 @@ const ifj = @import("ifj24.zig");
 pub fn main() void {
     ifj.write("yag er'ouy\n");
     
-    var a: f32 = 12.5;
+    var a: f64 = 12.5;
 
     a = "kys";  //Prirazeni []u8 do f64
